requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.1
requests[socks]==2.31.0
diff-match-patch==20230430
flask==3.03
gunicorn==23.0.0
